+++ mixins/ProfilesMixin.lua
@@ -6,7 +6,7 @@ MJProfilesMixin = util.createFromEventsMixin()
 
 
 function MJProfilesMixin:onLoad()
-	self.addonName = format("%s_ADDON_", strupper(addon))
+	self.addonName = ("%s_ADDON_"):format(addon:upper())
 	StaticPopupDialogs[self.addonName.."NEW_PROFILE"] = {
 		text = addon..": "..L["New profile"],
 		button1 = ACCEPT,
@@ -20,10 +20,11 @@ function MJProfilesMixin:onLoad()
 			local text = popup.editBox:GetText()
 			if text and text ~= "" then
 				if self.profiles[text] ~= nil then
-					StaticPopup_Show(self.addonName.."PROFILE_EXISTS")
+					self.lastProfileName = text
+					StaticPopup_Show(self.addonName.."PROFILE_EXISTS", nil, nil, data)
 					return
 				end
-				self.profiles[text] = data and MountsJournalUtil:copyTable(data) or {
+				self.profiles[text] = data and util:copyTable(data) or {
 					fly = {},
 					ground = {},
 					swimming = {},
@@ -44,10 +45,22 @@ function MJProfilesMixin:onLoad()
 			self.editBox:HighlightText()
 		end,
 	}
+	local function profileExistsAccept(popup, data)
+		if not popup then return end
+		local dialog = StaticPopup_Show(self.addonName.."NEW_PROFILE", nil, nil, data)
+		if dialog and self.lastProfileName then
+			dialog.editBox:SetText(self.lastProfileName)
+			dialog.editBox:HighlightText()
+			self.lastProfileName = nil
+		end
+	end
 	StaticPopupDialogs[self.addonName.."PROFILE_EXISTS"] = {
 		text = addon..": "..L["A profile with the same name exists."],
 		button1 = OKAY,
+		hideOnEscape = 1,
 		whileDead = 1,
+		OnAccept = profileExistsAccept,
+		OnCancel = profileExistsAccept,
 	}
 	StaticPopupDialogs[self.addonName.."DELETE_PROFILE"] = {
 		text = addon..": "..CONFIRM_COMPACT_UNIT_FRAME_PROFILE_DELETION,
@@ -62,7 +75,7 @@ function MJProfilesMixin:onLoad()
 	self.journal = MountsJournalFrame
 	self.profiles = self.mounts.profiles
 	self.charDB = self.mounts.charDB
-	self.profilesNames = {}
+	self.profileNames = {}
 	self:SetText(self.charDB.currentProfileName or DEFAULT)
 	UIDropDownMenu_Initialize(self.optionsMenu, self.menuInit, "MENU")
 end
@@ -90,11 +103,12 @@ end
 
 function MJProfilesMixin:deleteProfile(profileName)
 	StaticPopup_Show(self.addonName.."DELETE_PROFILE", NORMAL_FONT_COLOR_CODE..profileName..FONT_COLOR_CODE_CLOSE, nil, function()
+		self.profiles[profileName] = nil
 		if self.charDB.currentProfileName == profileName then
 			self:setProfile()
+		else
+			self.mounts:setDB()
 		end
-		self.profiles[profileName] = nil
-		self.mounts:setDB()
 	end)
 end
 
@@ -103,7 +117,7 @@ function MJProfilesMixin:setProfile(profileName)
 	if profileName == nil or self.profiles[profileName] then
 		self.charDB.currentProfileName = profileName
 		self:SetText(profileName or DEFAULT)
-		self:event("SET_PROFILE")
+		self:event("UPDATE_PROFILE", true)
 	end
 end
 
@@ -113,9 +127,31 @@ function MJProfilesMixin:menuInit(level)
 
 	local btn = self:GetParent()
 	local info = UIDropDownMenu_CreateInfo()
-	info.notCheckable = true
 
-	if UIDROPDOWNMENU_MENU_VALUE == "new" then -- NEW PROFLE
+	if UIDROPDOWNMENU_MENU_VALUE == "settings" then
+		if btn.charDB.currentProfileName == nil then return end
+		info.isNotRadio = true
+		info.keepShownOnClick = true
+
+		info.text = L["Pet binding from default profile"]
+		info.checked = function() return btn.journal.db.petListFromProfile end
+		info.func = function(_,_,_, checked)
+			btn.journal.db.petListFromProfile = checked and true or nil
+			btn:event("UPDATE_PROFILE")
+		end
+		UIDropDownMenu_AddButton(info, level)
+
+		info.text = L["Zones settings from default profile"]
+		info.checked = function() return btn.journal.db.zoneMountsFromProfile end
+		info.func = function(_,_,_, checked)
+			btn.journal.db.zoneMountsFromProfile = checked and true or nil
+			btn:event("UPDATE_PROFILE")
+		end
+		UIDropDownMenu_AddButton(info, level)
+
+	elseif UIDROPDOWNMENU_MENU_VALUE == "new" then -- NEW PROFLE
+		info.notCheckable = true
+
 		info.text = L["Create"]
 		info.func = function() btn:createProfile() end
 		UIDropDownMenu_AddButton(info, level)
@@ -125,6 +161,7 @@ function MJProfilesMixin:menuInit(level)
 		UIDropDownMenu_AddButton(info, level)
 
 	elseif UIDROPDOWNMENU_MENU_VALUE == "specialization" then -- SPECS
+		info.notCheckable = true
 		info.hasArrow = true
 		info.keepShownOnClick = true
 
@@ -135,11 +172,10 @@ function MJProfilesMixin:menuInit(level)
 		end
 
 	elseif type(UIDROPDOWNMENU_MENU_VALUE) == "number" then -- PROFILE BY SPEC
-		info.notCheckable = nil
 		info.keepShownOnClick = true
 		info.arg2 = UIDROPDOWNMENU_MENU_VALUE
 
-		if #btn.profilesNames > 20 then
+		if #btn.profileNames > 20 then
 			local searchFrame = util.getDropDownSearchFrame()
 
 			info.text = DEFAULT
@@ -152,7 +188,7 @@ function MJProfilesMixin:menuInit(level)
 			end
 			searchFrame:addButton(info)
 
-			for _, profileName in ipairs(btn.profilesNames) do
+			for _, profileName in ipairs(btn.profileNames) do
 				info.text = profileName
 				info.arg1 = profileName
 				info.checked = function(self)
@@ -178,7 +214,7 @@ function MJProfilesMixin:menuInit(level)
 			end
 			UIDropDownMenu_AddButton(info, level)
 
-			for _, profileName in ipairs(btn.profilesNames) do
+			for _, profileName in ipairs(btn.profileNames) do
 				info.text = profileName
 				info.arg1 = profileName
 				info.checked = function(self)
@@ -194,6 +230,7 @@ function MJProfilesMixin:menuInit(level)
 		end
 
 	else -- MENU
+		info.notCheckable = true
 		info.isTitle = true
 
 		info.text = L["Profiles"]
@@ -201,11 +238,11 @@ function MJProfilesMixin:menuInit(level)
 
 		UIDropDownMenu_AddSeparator(level)
 
-		info.isTitle = nil
 		info.notCheckable = nil
+		info.isTitle = nil
 		info.disabled = nil
 
-		if #btn.profilesNames > 20 then
+		if #btn.profileNames > 20 then
 			local searchFrame = util.getDropDownSearchFrame()
 
 			info.text = DEFAULT
@@ -213,7 +250,7 @@ function MJProfilesMixin:menuInit(level)
 			info.func = function() btn:setProfile() end
 			searchFrame:addButton(info)
 
-			for _, profileName in ipairs(btn.profilesNames) do
+			for _, profileName in ipairs(btn.profileNames) do
 				info.text = profileName
 				info.arg1 = profileName
 				info.checked = function(self) return btn.charDB.currentProfileName == self.arg1 end
@@ -229,7 +266,7 @@ function MJProfilesMixin:menuInit(level)
 			info.func = function() btn:setProfile() end
 			UIDropDownMenu_AddButton(info, level)
 
-			for _, profileName in ipairs(btn.profilesNames) do
+			for _, profileName in ipairs(btn.profileNames) do
 				local buttonFrame = util.getDropDownMenuButtonFrame()
 				buttonFrame.text = profileName
 				buttonFrame.arg1 = profileName
@@ -253,6 +290,12 @@ function MJProfilesMixin:menuInit(level)
 		info.notCheckable = true
 		info.hasArrow = true
 
+		if btn.charDB.currentProfileName ~= nil then
+			info.text = L["Profile settings"]
+			info.value = "settings"
+			UIDropDownMenu_AddButton(info, level)
+		end
+
 		info.text = L["New profile"]
 		info.value = "new"
 		UIDropDownMenu_AddButton(info, level)
@@ -275,8 +318,8 @@ end
 
 function MJProfilesMixin:onClick()
 	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
-	wipe(self.profilesNames)
-	for k in pairs(self.profiles) do tinsert(self.profilesNames, k) end
-	sort(self.profilesNames, function(a, b) return a < b end)
+	wipe(self.profileNames)
+	for k in pairs(self.profiles) do tinsert(self.profileNames, k) end
+	sort(self.profileNames, function(a, b) return a < b end)
 	ToggleDropDownMenu(1, nil, self.optionsMenu, self, 111, 15)
 end