+++ mixins/WorldMapMixin.lua
@@ -7,7 +7,6 @@ function MJMapCanvasMixin:onLoad()
 	self.highlight = self.child.HighlightTexture
 	self.detailLayerPool = CreateFramePool("FRAME", self.child, "MapCanvasDetailLayerTemplate")
 	self.explorationLayerPool = CreateTexturePool(self.child.Exploration, "ARTWORK", 0)
-	self.navBar:on("MAP_CHANGE", function() self:refresh() end)
 end
 
 
@@ -72,15 +71,19 @@ end
 
 
 function MJMapCanvasMixin:refresh()
-	if self:IsShown() then
-		self:refreshLayers()
-		self.navigation:Refresh()
-	end
+	self:refreshLayers()
+	self.navigation:Refresh()
 end
 
 
 function MJMapCanvasMixin:onShow()
 	self:refresh()
+	self.navBar:on("MAP_CHANGE.WORLDMAP", function() self:refresh() end)
+end
+
+
+function MJMapCanvasMixin:onHide()
+	self.navBar:off("MAP_CHANGE.WORLDMAP")
 end