+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-* Added search on existing lists.
+* Added cancelform to Stag and Treant form.
+* Added profile settings with options: "Pet binding from default profile", "Zones settings from default profile".
 * Minor fixes.